% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_treatment.R
\name{plot_hist}
\alias{plot_hist}
\title{Plot a histogram of variables}
\usage{
plot_hist(tmp, pfsdycr)
}
\arguments{
\item{tmp}{a dataframe with target variables}

\item{var}{the target variable}
}
\description{
TODO: add this later
}
