% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_treatment.R
\name{plot_panels}
\alias{plot_panels}
\title{Plot panels for tmp data}
\usage{
plot_panels(plot_data)
}
\arguments{
\item{plot_data}{a cleaned data with specific variables}
}
\description{
In hw1, this function is used to plot the correlations among all variables in tmp data
}
\examples{
tmp<-data(tmp);plot_panels(tmp)
}
