% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyze_survival_time.R
\name{plot_survival}
\alias{plot_survival}
\title{Plot a survival result}
\usage{
plot_survival(fit)
}
\arguments{
\item{fit}{The survival analysis model}
}
\value{
p a survival plot
}
\description{
plot a survival result between different group, to find whether different drugs have any different effect on survival time
}
\examples{
plot_survival(survival_fit(tmp))
}
