% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary_table.R
\name{summarize_table}
\alias{summarize_table}
\title{summarize our data}
\usage{
summarize_table(data_cleaned)
}
\arguments{
\item{data_cleaned}{a cleaned data with important variables (usually you can directly use the "dl" data from this package)}
}
\value{
a summary table
}
\description{
This function takes a cleaned dataframe containing important parameters, and return the summary table
}
\examples{
dl<-data(dl);summary_table(dl)

}
