% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_treatment.R
\name{plot_hist}
\alias{plot_hist}
\title{Plot a histogram}
\usage{
plot_hist(cleaned_data, pfsdycr = "pfsdycr", atrt = "atrt")
}
\arguments{
\item{cleaned_data}{a cleaned data with specific variables}

\item{pfsdycr}{default}

\item{atrt}{default}
}
\description{
Plot the histogram pf the variable named "pfsdycr", to find the correlation between treatment group and experiment group
}
\examples{
plot_hist(tmp)

}
