% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_treatment.R
\name{plot_trt}
\alias{plot_trt}
\title{Plot variables}
\usage{
plot_trt(plot_data, pfsdycr = "pfsdycr", dthdy = "dthdy", trt = "trt")
}
\arguments{
\item{plot_data}{a cleaned data with specific variables}

\item{pfsdycr}{default}

\item{dthdy}{default}

\item{trt}{default}
}
\description{
In hw1, this function is used to explore relationships between pfsdycr and dthdy
}
\examples{
plot_trt(tmp)

}
