% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyze_survival_time.R
\name{summarize_cox}
\alias{summarize_cox}
\title{print a cox result}
\usage{
summarize_cox(tmp_d, pfsdycr = "pfsdycr", pfscr = "pfscr", atrt = "atrt")
}
\arguments{
\item{tmp_d}{tmp data from this package}

\item{pfsdycr}{default}

\item{pfscr}{default}

\item{atrt}{default}
}
\value{
univariate cox fit model
}
\description{
print a cox result
}
\examples{
summarize_cox(tmp)
}
